Allow users to determine tables via python typing (#38)
* Allow users to determine tables via python typing

In order to release a safe enterprise SQL solution, we need the ability
for users to determine which values are identifiers, and which should be
entirely untrusted. Previously, we had people determine "safe parameters"
in the dag parameters, but this was very disruptive. With this new
system, users can simply use python typing to type something as a ``Table``

* fix typing

* nit

* fix snow