As a step towards "maturing" the astro DAG authoring project, we must rewrite our tests to ensure that every integration test runs against every database.
This step will simultaneously reduce the number of tests we need to maintain, make testing much simpler as we add new databases, and will make a future refactor much simpler as we can ensure proper coverage.

To do this, we will take advantage of two features in pytest, fixtures and parameterize.

For this ticket, we will update the `append` function.

**Acceptance criteria**
Have a single test file validating `append` across all databases.

1. Integration tests should be marked with `pytest.marker.integration`:

* Each test should work across all databases
* Validating both `Table` and `TempTable` as inputs
* Use `test_utils.run_dag`
* Use PyTest `fixtures` and `parameterize` to have a single main test that will validate transform across multiple databases

The tests should validate these scenarios:
1. appending two tables against a single column
2. appending two tables against multiple columns
3. appending against all fields by not specifying fields
4. appending with casting
5. append with some casted fields and some uncasted fields
6. test with two different databases (should fail)