aql.render can now use the dag's template_searchpath (#231)
Context
When instantiating Airflow DAGs, users can declare template_searchpath, which is a list of folders (non-relative) that define where jinja will look for templates. By default, it contains the DAGs folder.

Many SQL-like operators, such as the SnowflakeOperator and BigQueryOperator, support retrieving SQL files from this path, as described in the doc.

At the moment, Astro 0.5.1 expects users to declare the path to the SQL files as an argument of the aql.render method.