Enrich modules with TypeHints (part 2/9) (#296)
Solves partially #281

Add TypeHints to the following modules:

│   ├── settings.py
│   ├── sql
│   │   ├── __init__.py
│   │   ├── operators
│   │   │   ├── agnostic_aggregate_check.py
│   │   │   ├── agnostic_boolean_check.py

Also, refactor DependencyMissing to support TypeHint

We use the following in some parts of the code:
    Union[BigQueryHook, PostgresHook, SqliteHook, SnowflakeHook, MissingPackage]

Including in:
    astro/utils/__init__.py

However, MyPy was failing with the message:
    raise TypeError(f"{msg} Got {arg!r:.100}.")
https://github.com/python/cpython/blob/main/Lib/typing.py#L162-L192

While MissingPackage was being used as an object, the other return items
would be classes. Now everything is returned as a class.