Run tests against multiple airflow versions (#370)
We want to make sure that Astro SDK continues working with Airflow 2.2.5 and Airflow 2.3 - so this commit adds this capability in nox.

A problem I foresee is some tests failing intermittently as some of our "integration tests" don't have unique ids for tables, dataset which means there might be times when an integration test is running in 2.2.5 which deletes a resource that is being used by same test running for 2.3.

One "quick" solution is to run them sequentially but that will increase total test time. A better solution will be to fix the tests but that will take time.