Refactor `MergeOperator` (#422)
This PR refactors the Merge operator to use the new interfaces. This PR intends to keep the same logic as now and just use new interfaces so we can delete all the old files from old interfaces. The scope to whether or not we should have a single function or not is kept for https://github.com/astronomer/astro-sdk/pull/383 

part of https://github.com/astronomer/astro-sdk/issues/343