Refactor `AppendOperator`
This PR refactors the Append operator to use the new interfaces.

Breaking Changes:
- Removed casting functionality
- changed the interface from which uses `source_to_target_columns_map` instead of `columns`:

```
    def append(
        self,
        main_table: Table,
        columns: List[str],
        casted_columns: dict,
        append_table: Table,
    ):
```

to

```
def append(
    source_table: Table,
    target_table: Table,
    source_to_target_columns_map: Optional[Dict[str, str]] = None,
    **kwargs,
):
```

This PR intends to keep the same logic as now and just use new interfaces so we can delete all the old files from old interfaces. The scope to whether or not we should have a single function or not is kept for https://github.com/astronomer/astro-sdk/pull/383

closes #343
closes https://github.com/astronomer/astro-sdk/issues/335

Co-Authored-By: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>
Co-Authored-By: Utkarsh Sharma <utkarsharma2@gmail.com>