POC of creating a table cleanup task
This POC proves that we can extend the Airflow DAG object to create a
task that lives downstream of all other tasks, and that said task can
access the output of all previous tasks. We can later further optimize
this to ensure that it only ever pulls table objects (based on the task
type).