Create scripts to launch benchmark testing to Google cloud (#496)
Currently, users are running benchmark tests on their local machines. This is a) inconsistent as much of the time might be tied to network time and b) bad as it discourages consistently running benchmarks (E.g. as a daily job or as a part of CI).

Now users will have the ability to launch a GKE cluster, push an image for the Astro SDK, and launch the job to run on GKE

closes: #432

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>