Get configs via Airflow Configuration manager (#505)
What is the current behavior?

Currently, we are reading the config info. directly from the env variables using - os.get(). A better way to read the values is from airflow configs.

closes: #503
What is the new behavior?

We can use the below code to read from the airflow's config. By assuming a section in the airflow's config named - astro

Does this introduce a breaking change?

No