Load files from GCS to Bigquery using BigqueryHook (#489)
What is the current behavior?

When we run the load_file operator for a table in Bigquery and a file in GCS, the data is first loaded to local from the GCS and then uploaded in Bigquery. Because of this, we require both memory and network bandwidth on the worker node.

partially closes: #429
What is the new behavior?

We have used an optimized path that can directly ingest data from GCS to Bigquery, therefore removing both memory and bandwidth requirements on a worker node. Because of this, the speed and amount of data transfer increase.
Does this introduce a breaking change?

No