Rename `BaseSQLOperator` class (#518)
# Description
## What is the current behavior?
Currently the base class for all SQL Decorators is called "BaseSQLOperator". This is a misnomer, as this class makes a lot of assumptions around being a function decorator. Future contributors might be confused by this name and try to use it on non-decorator classes.


## What is the new behavior?
Change the name to the more accurate "BaseSQLDecoratorOperator" to ensure that users would only extend this class for creating decorators.

## Does this introduce a breaking change?

No