Allow running tests on PRs from forks + label (#546)
This will allow running tests on PRs from forks -- allowing outside contributors to contribute to this Repo. Committers of this repo (the one with write access) will need to add a "safe to test" label to run the tests. Until then the tests will not run.

This is deliberate as we want to be ultra sure on who can run CI with necessary secrets

closes https://github.com/astronomer/astro-sdk/issues/179

References: https://github.blog/2020-08-03-github-actions-improvements-for-fork-and-pull-request-workflows/