Create a `transform_file` operator (#677)
Currently there is no clean way to run `aql.transform` with a SQL file instead of a python function. This is highly limiting as SQL developers would not want to craft complex queries inside of a python function.

## What is the new behavior?

The new behavior is that we will now have an `aql.transform_file` function that can be called at runtime in the DAG. This is not a decorator, but is instead a python function that returns a task (similar to operators like `drop` and `merge`).

Instead of using function parameters, users would instead insert tables via the `parameters` kwarg.

It would look like this

```python
import pathlib

CWD = pathlib.Path(__file__).parent

@aql.dataframe
def validate(df: pandas.DataFrame):
    assert df.columns.tolist() == ["title", "rating"]

with dag:
    target_table = Table(name="test_is_{{ ds_nodash }}", conn_id="sqlite_default")
    table_from_query = aql.transform_file(
        file_path=str(pathlib.Path(CWD).parents[0]) + "/transform/test.sql",
        parameters={"input_table": imdb_table, "output_table": target_table},
    )
    validate(table_from_query)
```

This function also offers a nice "less magic" endpoint for astro-sdk users who might not yet be comfortable with pytho decorators.