Create subprojects for astro-sdk-python and astro-sql-cli (#744)
# Description
## What is the current behavior?
Currently, the astro-sdk project is set up only to handle one project. This has been fine when our only project was the Astro Python SDK, but now that we are expanding into more projects it is critical that we can run multiple projects in the same repo.


## What is the new behavior?
This PR takes all existing python-sdk work and puts it in a `python-sdk` directory. It then creates a `sql-cli` directory to ensure that we have a second working project (this process will make it easy to repeat across any other projects we want to add).

## Does this introduce a breaking change?

No