Map `File` & `Table` classes to Dataset (AIP-48)
Part of https://github.com/astronomer/astro-sdk/issues/611

This PR will map `File` and `Table` classes to inherit from `Dataset` object (if it is available - will be released in Airflow 2.4). This will allow users to schedule DAGs on a `File` and `Table` object when used with Airflow 2.4 and above.

As a follow-up to this PR, I will create a PR to automatically add "inlets" and "outlets" to all of the operators in SDK so that users can see the datasets that are produced or consumed by tasks.

Fix doc build error

Import from `attr` instead of `attrs`

Looks like Airflow has a limit set for 2.3.3 - https://github.com/apache/airflow/blob/2.3.3/setup.cfg#L88-L89 but not for 2.3.4.

This was fixed in Airflow 2.3.4 but I will change it to using attr.

`attrs` is a new namespace while `attr` is the old but working one