Avoid loading whole file into memory with load_operator for schema detection
We're loading whole file into memory and converting it into a byte stream
for schema auto-detection. As a result, jobs with large files e.g. 5GB and
higher are getting killed.
Opening this for discussion if we really need to load the whole file into
memory and if not needed just return the stream so that pandas dataframe
can fetch only the initial rows of a file for schema autodetection.