Create `upstream_tasks` parameter for dependencies independent of data transfers (#585)
# Description
## What is the current behavior?
Currently there is "native" way to set dependencies between tasks without passing either a table or a dataframe between said tasks. This is especially apparent when chaining raw_sql functions such as here: 

https://github.com/astronomer/astro-sdk/issues/569

In order to have a solution that works we need somehting that feels pythonic and doesn't rely on traditional airflow `>>` operators

closes: #569 


## What is the new behavior?

This PR adds in the ability to set dependencies using a parameter `upstream_tasks`. This parameter can be used at runtime to set deps without passing data. Notably we don't want to actually render the data (imagine a use-case where a task relies on 10 dataframe tasks)

## Does this introduce a breaking change?

No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary
- [ ] Add documentation around `upstream_tasks`
- [ ] solution should not render the data since the data is never used.