Avoid loading whole file into memory with load_operator for schema detection (#805)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

We're loading whole file into memory and converting it into a byte
stream for schema auto-detection. As a result, jobs with large files
e.g. 5GB and higher are getting killed.
Opening this for discussion if we really need to load the whole file
into memory and if not needed just return the stream so that pandas
dataframe can fetch only the initial rows of a file for schema
autodetection.

closes: #803 
closes: #826 
## What is the new behavior?
1. Do not load the whole file into memory but just return the smart_open
stream to be used by the dataframe.

2. Fix ndjson flattening logic.

## Does this introduce a breaking change?

No. Tested with postgres & Redshift.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Utkarsh Sharma <utkarsharma2@gmail.com>
Co-authored-by: Phani Kumar <94376113+phanikumv@users.noreply.github.com>