Allow users to disable auto addition of inlets/outlets via airflow.cfg
Astro SDK automatically adds inlets and outlets for all the operators if DATASET is supported (Airflow >=2.4).

Users can over-ride it on a task level by adding inlets and outlets themselves. However, for users who want to use SDK with Airflow 2.4 and above but do not want to leverage Data-aware scheduling, auto-addition of inlets/outlets would be annoying so this new config setting will provide an escape hatch for such users.

closes https://github.com/astronomer/astro-sdk/issues/857