Create a custom xcom backend that removes XCOM_PICKLING argument (#970)
# Description
## What is the current behavior?
Currenlty users need to set "xcom_pickling" to True to use the astro python SDK. We've received feedback from users that they do not want to set this variable and for some this is a blocking change.

closes: [AIRFLOW__CORE__ENABLE_XCOM_PICKLING](https://github.com/astronomer/astro-sdk/issues/795)


## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by this PR. -->

The new behavior is we now handle our own serialization and deserialization to ensure that our system works with the json storage option for Airflow. We do this by wrapping the existing xcom_backend (be that base or custom) and serializing before/deserializing after. This prevents the change from being breaking as users do not need to install a custom backend.

## Does this introduce a breaking change?
No

### Checklist
- [x] Created tests which fail without the change (if possible)
- [x] Extended the README / documentation, if necessary

Co-authored-by: Daniel Imberman <danielimberman@Daniels-MacBook-Pro.local>
Co-authored-by: Kaxil Naik <kaxilnaik@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>