Add DAG Renderer for run command (#1140)
# Description
## What is the current behavior?
Currently, if a user wants to run a DAG, we first must generate a DAG that generates each task, it's dependencies, and it's variables. This approach leads a DAG that is more explicit, but also can be hard to reach and involves a lot of code maintenance on our part/very little flexibility (e.g. if a user later wants to add a task, they either need to manually add it into the DAG or re-generate from scratch).

[#ISSUE](https://github.com/astronomer/astro-sdk/issues/946)

## What is the new behavior?

To explore multiple options, we will also create a DAG "render" option that does not generate any files, but instead creates a DAG object and runs that object. In the coming weeks we will compare the two solutions.

## Does this introduce a breaking change?
No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Felix Uellendall <feluelle@users.noreply.github.com>
Co-authored-by: Kaxil Naik <kaxilnaik@gmail.com>