Fix database setting in PostgresDatabase (#1276)
# Description

## What is the current behavior?

Currently, the CI is failing for postgres related tests, because the
postgres database is set to the table schema.

**More details:**

In
https://github.com/astronomer/astro-sdk/commit/313f6da5fc572a8a35c6fa3db90373a7586be4de#diff-9fcbc9dc6fc68bf0a0d62a485508779f5b1f6d749a197e0f1ab1758983d1eb1e
we added `kwargs.update({"database"` but PostgresHook did not support
`database` kwarg at that time. This is why it basically did not do
anything. Now after https://github.com/apache/airflow/pull/26744 got
released in the latest provider, it is being picked up and it shows that
we wrongly set `{"database": self.table.metadata.schema}`

## What is the new behavior?

We correctly set the database to the table's database.

## Does this introduce a breaking change?

No.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary