Use sqlite conn for load file workflow example
Previously, we were using bigquery conn to load data from GCS to
Bigquery. With #1282 we aim to keep it simple by not having to
create complex tests for using our example load workflow.
The PR replaces the existing load file workflow to load data from
a github file to load data into sqlite database.

We additionally are fixing minor issues where the File and Table
deserializers are pased void data values for their optional keys
so that the deserializer does not raise KeyError.