OpenLineage: Rename the name and namespace along with URI (#1306)
Describe the bug
- There have been new conventions introduced:
[Naming.md#local-file-system](https://github.com/OpenLineage/OpenLineage/blob/main/spec/Naming.md#local-file-system)
I believe sqlite counts as local file, therefore you could adjust naming
convention in astro as well.

A word of explanation: both IP and port are mandatory. It means that if
port is unknown (e. g. if file exists in Airflow's local file system
only) you should default to some value. In one of OL extractors it is
defaulted to paramiko.config.SSH_PORT, in most cases it's equal to 22.

Reference:
https://github.com/astronomer/astro-sdk/pull/1142#issuecomment-1321563229
- Add openlineage_dataset_uri() method to create URI as per each
database.

closes #1281

Co-authored-by: Ankit Chaurasia <8670962+sunank200@users.noreply.github.com>