Support Google Drive as file location (#1322)
This PR contains support for google drive as a file location. A file(s)
kept in google drive can be now loaded into the databases and can
perform operations on it.
This mostly just uses the Google API client to connect and retrieve the
file.

One endpoint is used:

*
[`files().list()`](https://developers.google.com/drive/api/v3/reference/files/list)
to find files.

To use those file APIs you need to kind of visualise traversing the
folders in a GUI. To get a file by path, you need to see what items are
available in a folder and select the one with the name you want. If the
item is a folder, you need to run the list against it again to see what
items are in it. Not very efficient, but there are not many choices.


## Does this introduce a breaking change?

No.


### Checklist
- [x] Created tests which fail without the change (if possible)

Close #1044

Co-authored-by: uranusjr <uranusjr@gmail.com>