Add basic spark utilities that will provide the basis of future spark functions (#1320)
# Description
## What does this PR do?

Creating Databricks support is going to be a fairly large task, and we
want to break up PRs into smaller pieces wherever possible. This PR lays
down the initial framework for databricks support. It only offers
"transform" functionality, but it creates the Delta database class,
CI/CD, and necessary imports

closes: https://github.com/astronomer/astro-sdk/issues/1328

## Does this introduce a breaking change?
No, though it's currently questionable whether this system will be
compatible with Airflow 2.2.5.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary