`LoadFile` should return a dataframe if not using XCom backend
Currently, the `LoadFileOperator` converts dataframe to a File object when returning. However, this should be handled by the Custom XCom backend and not in `LoadFileOperator`. Otherwise, we won't be consistent with dataframe and transform decorators who, too returns dataframes.

From Airflow 2.5 and above, this should be handled automatically, for older versions, users should use our Custom XCom backend so that their XCom table isn't fully of GBs of data.