Improve airflow configuration (#1350)
# Description

## What is the current behavior?

Currently, our airflow configuration consists of settings in the config
file as well as in env vars. This PR addresses it by using a single
approach.

closes: https://github.com/astronomer/astro-sdk/issues/1219

## What is the new behavior?

We only use env vars for setting airflow configuration. Unfortunately
every time we set those, we need to reload the modules for them to read
the updated env var.

## Does this introduce a breaking change?

No.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary