`LoadFile` should return a dataframe if not using XCom backend (#1348)
Currently, the `LoadFileOperator` converts dataframe to a File object
when returning. However, this should be handled by the Custom XCom
backend and not in `LoadFileOperator`. Otherwise, we won't be consistent
with dataframe and transform decorators who, too returns dataframes.

From Airflow 2.5 and above, this should be handled automatically, for
older versions, users should use our Custom XCom backend so that their
XCom table isn't fully of GBs of data.

part of https://github.com/astronomer/astro-sdk/issues/1337

## Does this introduce a breaking change?
No, if you are using our XCom backend. Yes, if you are not! but this
fixes the behaviour and makes it consistent with other decorators and
operators that return dataframes

### Checklist
- [x] Created tests which fail without the change (if possible)
- [x] Extended the README / documentation, if necessary

(cherry picked from commit 128339470f8348fa9df88e9d9fe7e4562753aa44)