Add working examples and tests for all supported connection types (#1388)
# Description

## What is the current behavior?

Currently, we don't have any working example(s) of bigquery connections,
because we thought the `extra` is broken. But apparently we did not
configure it right.

**Prior to Airflow 2.3**, the `extra` has to be of type string and also
in case of bigquery all of the extras needed to be prefixed with
`extra__google_cloud_platform__`. Another example is snowflake where the
`extra` needed to be a json as well to work for airflow 2.2.5 which we
support.

closes: #1277

## What is the new behavior?

We add examples and tests for all currently supported connection types.
Moreover configure the ci pipeline to be able to test them.

## Does this introduce a breaking change?

No.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary