Create .env file on project initialisation + important refactorings (read desc) (#1391)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
Currently, there is no way for Astro CLI users to specify environment
variables that can be used for configuring connections as we launch
docker containers from Astro CLI commands which do not have access
to host environment variables

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->
related: #1329 



## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->
As part of project initialisation, we create a .env file in the
initialised project and
ask the user to configure environment variables in this file. Astro CLI
will mount
this .env file in the docker containers so that those environment
variables are
available in the docker container since we use `dotenv` library to load 
environment variables from .env file. Such environment variables can be
used
to expand yaml files to create connections including secrets as
environment variables.
e.g.
https://legacy.docs.greatexpectations.io/en/latest/guides/how_to_guides/configuring_data_contexts/how_to_use_a_yaml_file_or_environment_variables_to_populate_credentials.html

Remove transform env specific config implementation and move the
needed relevant connection transformation into the convert_to_connection
method.

Remove the method to write config to yaml as it overwrites environment
vars placeholder with the expanded vars when read into config instance.

## Does this introduce a breaking change?


### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary