Redirect default airflow_home to tmp directory in tests (#1384)
# Description

## What is the current behavior?

Currently, when testing with different versions of airflow locally,
sometimes airflow still writes to the default airflow home directory.
However, we are not using the default.

**Note:**
We are setting the `AIRFLOW_HOME` on project init, but not before that.
Therefore any airflow import before that will write to the default
airflow home directory.

## What is the new behavior?

Set the default airflow home to a new tmp directory. So that any call
before calling project init writes to the tmp directory.

We also migrate all installs to explicitly use `poetry run` to ensure we
are always installing to the correct venv. And we move typer env vars
from workflow file to noxfile to be able to run nox locally with the
same required env vars.

## Does this introduce a breaking change?

No.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary