Fix elapsed seconds for run command (#1406)
The previous implementation just considers the microseconds and ignores
all other attributes like seconds, minutes and hours. e.g if the
difference is 00:00:05.7878, it only took 7878 and wrote in the output
as 0.78 seconds, whereas the output should have been 5.79 seconds with
rounding the output to 2 decimal places. The fix includes using the
total_seconds() method of the timedelta object to calculate the dag run
duration accurately for the output.