Move file config/global/configuration.yml -> config/global.yml
Based on previous discussions and observations from users, it seems
that having a global directory creates confusion because global
itself is not any environment and the directories within config
directory are meant to be environments. So, we are no longer creating
the global directory as part of the initialised project and moving
the global configuration to top level global.yml file in the config
directory.