Create region specific datasets for bigquery (#1416)
Currently, we are assuming all the bigquery datasets to be in the
default region - `US`
If the user tries to access a preexisting table in a different region
say, `europe-west6` all the Temp tables will try to be created in `US`
the region, which will result in the error.

closes: #1369

1. The solution is to create `tmp_astro` region specific. But we cannot
do that since `<project_id>.<dataset_id>` is unique. So we are creating
a unique dataset id like - `<dataset_id>__<region_id>`, this region_id.
2.  Added a default setting to select the Bigquery default region.
3.  Added `location` attribute to Metadata class.

Nope

Co-authored-by: Phani Kumar <94376113+phanikumv@users.noreply.github.com>
Co-authored-by: Felix Uellendall <feluelle@users.noreply.github.com>
(cherry picked from commit 730d273b32e4f6a453f90837942a8490bc766cf9)