Add load to spark gcs (#1404)
# Description

## What is the new behavior?
This PR adds the ability for GCP users to load GCS files to delta for
processing in databricks automatically. It is worth noting that this is
optional, as users can turn off secret loading and manually add GCS
settings to their databricks cluster using [this
guide](https://docs.gcp.databricks.com/external-data/gcs.html) if they
prefer. The main benefit here is that users do not need to set these
settings ahead of time.

## Does this introduce a breaking change?
No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary