Add merge function to databricks delta (#1476)
# Description
## What is the current behavior?

Currently, the merge function returns a `NotImplementedError` if a user
attempts to merge two delta tables. This leaves the delta astro SDK
experience incomplete and can create confusion for users.

Closes #1451 

## What is the new behavior?
The new behavior is to add a merge function in a similar fashion to how
we generate a merge function in snowflake. We have found these two
interaces to be fairly similar so it might be worth looking into whether
the functions can be merged in a future ticket.

## Does this introduce a breaking change?
No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary