Move databricks module into the databases directory (#1555)
## What is the current behavior?
Currently the "databricks" module lives at the top level of the python
SDK, this can be somewhat confusing as it might lead users to think
there is non-database functionality in the databricks module. By moving
the databricks module into the database module, we insulate the class
and ensure that it is treated like any other database going forward.