Change load_options to list[LoadOptions] from LoadOptions (#1540)
# Description
## What is the current behavior?
In load_file() we use multiple integrations with Pandas, and Databases
at the same time. But We currently just have one parameter to pass in
integration-specific parameters.

closes: https://github.com/astronomer/astro-sdk/issues/1532

## What is the new behavior?
We have changed load_options params to List[LoadOptions] so that we can
pass multiple integration parameters in a list.
Passing multiple values to `load_options` leads to the problem of
mapping. How do we know that PandasLoadOptions need to be passed to
dataframe specific methods?

We solved the mapping problem by having integration classes having
`LOAD_OPTIONS_CLASS_NAME` parameter to store the `LoadOptions` subclass
for that integration and initialize the File-type, File-location, and
Database objects with the right LoadOption class.

## Does this introduce a breaking change?
No, since we never released the `load_options`

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary