Add error message for validate connections (#1554)
# Description

## What is the current behavior?

Currently, we are not giving any hints to the user about an
invalid/failed connection.

closes: #1535

## What is the new behavior?

Print an error message when validating a connection.

## Does this introduce a breaking change?

Yes, kind of. We are not erroring out anymore when the file path
internally could not get resolved in case of sqlite connections. This is
so that an invalid sqlite connection does not block you from running any
generate/validate/run commands. Previously it was raising an exception
which led to an immediate exit.

### Checklist
- [x] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>