Add documentation around LoadOptions with example DAGs (#1567)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
In Astro SDK 1.4 we introduce a series of LoadOptions classes for
classes such as databricks, pandas, and snowflake. Each of these classes
is unique and we should create documentation that best showcases how
users can use these classes to create the loading settings they want.

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->
closes: #1556


## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Add example DAG using `PandasLoadOptions`, `SnowflakeLoadOptions` and
`DeltaLoadOptions`
- Add the documentation for `PandasLoadOptions`, `SnowflakeLoadOptions`
and `DeltaLoadOptions`. Add detailed documentation of how a user can use
these classes.
- Add examples for PandasOptions, DeltaOptions, and Snowflake Options.

## Does this introduce a breaking change?
No

### Checklist
- [x] Created tests which fail without the change (if possible)
- [x] Extended the README/documentation, if necessary