Disable logging for non-debug mode (#1569)
# Description

## What is the current behavior?

Currently, we oftentimes print some logging statements eventhough we are
not using the verbose nor debug flag.

## What is the new behavior?

We disable logging for non-debug mode. The only exception is when
verbose mode is set it will selectively enable logging. If debug mode is
set enable logging for a selective list of loggers.

## Does this introduce a breaking change?

Not really. There are really no logs being shown when non-debug and
non-verbose mode is used.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

Co-authored-by: Tatiana Al-Chueyr <tatiana.alchueyr@gmail.com>