move load_file and example_dag tests into their own section (#1588)
# Description

In trying to optimize the astro SDK CI system, I find that the load_file
and example_dag sections are two sections that just can NOT be optimized
using the mocking library. The load_file library can not be optimized
because we are literally testing whether the files are being properly
loaded into the external databases, and the example DAGs need to run
exactly to spec so we are sure users can run them as-is.

To handle this limitation while still bringing down the total CI time,
we separate the load_file tests and example_dags tests into their own
section where we can split them using pytest-split and run them in
parallel (where before they would run sequentially on a single worker).

We can now run all load_file tests within 10 minutes across three
workers. and example_dag tests in 12 minutes.

## Does this introduce a breaking change?


### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary